src (rename): useStore into useContextHook
