+++ doc-projet-en/design/navigation-panning.md
@@ -29,6 +29,7 @@ Time estimate: 4h
 2020-01-16T00:09+01 -> 2020-01-16T00:22+01 :: moving the display block to the top (.7x the expected time)
 2020-01-16T00:23+01 -> 2020-01-16T00:30+01 :: changing the button symbols (.7x the expected time)
 2020-01-16T07:13+01 -> 2020-01-16T07:45+01 :: disabling arrow that cannot be used (1.5x the expected time)
+2020-01-16T18:35+01 :: add a context for accessing the display (15 minutes)
 :: Move key logic to display.tsx
 
 #### Todo
