+++ src/ui/control/Controller.tsx
@@ -12,7 +12,7 @@ import * as React from 'react'
 import { Rule } from '../editor/components/Rule'
 import { useSharedStyle } from '../style'
 import { clx } from '../util/clx'
-import { useStore } from '../util/useStore'
+import { useStore } from '../util/useContextHook'
 import { AlignTime } from './components/AlignTime'
 import { CanvasHeight } from './components/CanvasHeight'
 import { CanvasWidth } from './components/CanvasWidth'
