+++ src/ui/control/components/CaSizeSelector.tsx
@@ -1,7 +1,7 @@
 import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import { errorCheck } from '../../../util/errorCheck'
-import { useStore } from '../../util/useStore'
+import { useStore } from '../../util/useContextHook'
 import { SelectorInput } from './SelectorInput'
 
 let validation = (value: string) => {
