+++ src/ui/control/components/FieldPosT.tsx
@@ -1,7 +1,7 @@
 import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import { errorCheck } from '../../../util/errorCheck'
-import { useStore } from '../../util/useStore'
+import { useStore } from '../../util/useContextHook'
 import { SelectorInput } from './SelectorInput'
 
 export let FieldPosT = observer(() => {
