+++ src/ui/control/components/panning/HorizontalPanning.tsx
@@ -2,7 +2,7 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core'
 import Button from '@material-ui/core/Button'
 import ButtonGroup from '@material-ui/core/ButtonGroup'
 import * as React from 'react'
-import { useStore } from '../../../util/useStore'
+import { useStore } from '../../../util/useContextHook'
 import { Xelement } from '../../../util/Xelement'
 import { observer } from 'mobx-react-lite'
