+++ src/ui/editor/Editor.tsx
@@ -11,7 +11,7 @@ import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import { useSharedStyle } from '../style'
 import { clx } from '../util/clx'
-import { useStore } from '../util/useStore'
+import { useStore } from '../util/useContextHook'
 import { PropertyList } from './components/PropertyList'
 import { Rule } from './components/Rule'
 import { RuleEditor } from './components/RuleEditor'
