+++ src/www/global.ts
@@ -1,9 +1,11 @@
 import { createContext } from 'react'
 import { Hub } from '../state/hub'
 import { Store } from '../state/store'
+import { Display } from '../display/display'
 
 export let storeContext = createContext<Store | null>(null)
 export let hubContext = createContext<Hub | null>(null)
+export let displayContext = createContext<Display | null>(null)
 
 export let timeStore: {
    [k: string]: {
